Solve memory leak for documents with nested hash
Without this patch, BSON::XS has a memory leak in the C code that is
running always when document has subdocument, i.e. contains a nested
hash (or IxHash):

```
my $hash = {
    'nested' => {
        'key' => 'str',
    }
};

```

The following C structure is currently used to trace circular
references on hashes:

```
 typedef struct _stackette {
   void *ptr;
   struct _stackette *prev;
 } stackette;
```

New stackette is allocated for every document and its subdocuments.

But to free this structure it is not enough to call `Safefree(stack);`
on first stackette, because all next stackettes remain not free and
their addresses get lost.

I tried to code less and came up with the solution to improve stackette
structure to keep all next references that can be used to free next
stackettes in right place.